requirements.txt
discord.py>=2.3.2
wavelink==3.2.0
python-dotenv>=1.0.0
PyNaCl>=1.4.0
